requirements.txt
future==0.18.2
mysql-connector-python==8.0.21
protobuf==3.13.0
PySide2==5.15.0
PySimpleGUIQt==0.35.0
shiboken2==5.15.0
six==1.15.0
treelib==1.6.1
